% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct.evidence.plot.R
\name{direct.evidence.plot}
\alias{direct.evidence.plot}
\title{Plot for direct evidence proportions in a network meta-analysis using \code{netmeta}}
\usage{
direct.evidence.plot(x, random=FALSE, comparison.label.size=2,
numeric.label.size=3, subplot.ratio=c(5, 1.3, 1.3))
}
\arguments{
\item{x}{An object of class \code{netmeta} containing the results of a network meta-analysis
using the \code{\link[netmeta]{netmeta}} function.}

\item{random}{Logical. If set to \code{TRUE}, results for the random-effects model are displayed.
If set to \code{FALSE}, results for the fixed-effect model are displayed. \code{FALSE} by default.}

\item{comparison.label.size}{A numeric value for the size of comparison labels
to be used in the plot. Default is \code{2}.}

\item{numeric.label.size}{A numeric value for the label size of numeric values
to be used in the plot. Default is \code{3}.}

\item{subplot.ratio}{A numeric vector containing three numbers. Defines the width for each of
the three subplots included in the plot (from left to right). Default is \code{c(5,1.3,1.3)}.}
}
\value{
\itemize{
\item \code{data}: A \code{data.frame} containing columns for the proportion of direct and indirect
evidence of each comparison (\code{proportion.direct} and \code{proportion.indirect}), the
mean path length (\code{meanpath}) and the minimal parallelism (\code{minpar})
for each comparison.
\item \code{plot}: The generated plot (if the function output was saved to an object).
}
}
\description{
This function plots relevant measures quantifying the direct evidence proportion, mean path length
and aggregated minimal parallelism of a frequentist network meta-analysis model generated by
\code{\link[netmeta]{netmeta}}.
}
\details{
The function generates a plot containing three subplots displaying relevant characteristics
to evaluate the reliability of effect size estimates within a network meta-analysis model.
\itemize{
\item \strong{Direct Evidence Proportion}. This bar chart displays the proportion of direct
evidence (orange) contained in each network estimate. It is of note that both direct and indirect
evidence may contribute to the violation of the assumption of consistency underlying network
meta-analysis models. Nevertheless, this plot allows to distinguish comparison estimates
 for which direct evidence was used, and to what extent, and comparisons which had to be inferred
 by indirect evidence alone.
\item \strong{Minimal Parallelism}. This bar chart displays the minimum number of independent paths
contributing to the effect estimate on an aggregated level. Large values of parallelism can be
interpreted as supporting the robustness of the estimate.
\item \strong{Mean Path Length}. This bar chart displays the mean path length, which characterizes
the degree of indirectness of an estimate. Higher mean path lengths indicate less reliable
estimates, given that more similarity assumptions have to be made when serially combining
direct comparisons. Following König, Krahn and Binder
(\href{https://www.ncbi.nlm.nih.gov/pubmed/24123165}{2013}), comparisons with mean path lengths
greater than two should be interpreted with caution. This threshold is displayed as a blue vertical
line in the plot.
}
}
\examples{
# Load Senn2013 data from netmeta
suppressPackageStartupMessages(library(netmeta))
data(Senn2013)

# Conduct network meta-analysis (fixed-effects model)
nma = netmeta(TE, seTE, treat1, treat2, studlab,
              data=Senn2013, sm='MD', comb.random=FALSE)

# Generate the plot
dep = direct.evidence.plot(nma, random=FALSE, comparison.label.size = 1,
numeric.label.size=1, subplot.ratio=c(3,1,1))
dep
}
\references{
Harrer, M., Cuijpers, P., Furukawa, T.A, & Ebert, D. D. (2019).
\emph{Doing Meta-Analysis in R: A Hands-on Guide}. DOI: 10.5281/zenodo.2551803. \href{https://bookdown.org/MathiasHarrer/Doing_Meta_Analysis_in_R/frequentist-network-meta-analysis.html}{Chapter 11.1}

König J., Krahn U., Binder H. (2013): Visualizing the flow of evidence in network meta-analysis and
characterizing mixed treatment comparisons. \emph{Statistics in Medicine, 32}, 5414–29
}
\seealso{
\code{\link[netmeta]{netmeta}}, \code{\link[netmeta]{netmeasures}}
}
\author{
Mathias Harrer & David Daniel Ebert
}
