% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmetar.R
\docType{package}
\name{dmetar}
\alias{dmetar}
\alias{dmetar-package}
\title{dmetar: Companion R package for the guide 'Doing Meta-Analysis in R'}
\description{
\code{dmetar} serves as the companion R package for the
guide \href{https://bookdown.org/MathiasHarrer/Doing_Meta_Analysis_in_R/}{Doing Meta-Analysis in R}
by Mathias Harrer, Pim Cuijpers, Toshi Furukawa and David Daniel Ebert.
}
\details{
The package contains complementary functions to facilitate the conduction of meta-analyses
using the \pkg{meta}, \pkg{metafor}, \pkg{netmeta} and \pkg{gemtc} packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MathiasHarrer/dmetar}
}

}
\author{
Mathias Harrer \email{mathias.harrer@fau.de}, David Daniel Ebert \email{david.daniel.ebert@gmail.com}
}
