% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggers.test.R
\name{eggers.test}
\alias{eggers.test}
\title{Perform Egger's test of the intercept}
\usage{
eggers.test(x)
}
\arguments{
\item{x}{An object of class \code{meta}, generated by the \code{metabin}, \code{metagen},
\code{metacont}, \code{metacor}, \code{metainc}, or \code{metaprop} function.}
}
\value{
Returns a data.frame containing the following columns:
\itemize{
\item \code{Intercept}: The intercept (bias).
\item \code{ConfidenceInterval}: The 95\% confidence interval of the intercept.
\item \code{t}: The t-statistic for the intercept test.
\item \code{p}: The p-value for Egger's test.
}
}
\description{
This function performs Egger's test of the intercept for funnel plot asymmetry using an object
of class \code{meta}.
}
\details{
Performs Egger's test (Egger et al., 1997) for funnel plot asymmetry.
The \code{\link[meta]{metabias}} function is called internally. Egger's test may lack
the statistical power to detect bias when the number of studies is small. Sterne et al.
(2011) recommend to perform funnel plot asymmetry tests only when \eqn{k \geq 10}. A warning
is therefore printed when the number of studies in the \code{meta} object is \eqn{k < 10}.
}
\examples{
# Create meta-analysis results using the 'metagen' function
suppressPackageStartupMessages(library(meta))
data(ThirdWave)
m = metagen(TE, seTE, studlab = paste(Author),
    data = ThirdWave, comb.random = FALSE, hakn=TRUE)

# Plug result into 'eggers_test' function
eggers.test(m)
}
\references{
Harrer, M., Cuijpers, P., Furukawa, T.A, & Ebert, D. D. (2019).
\emph{Doing Meta-Analysis in R: A Hands-on Guide}. DOI: 10.5281/zenodo.2551803. \href{https://bookdown.org/MathiasHarrer/Doing_Meta_Analysis_in_R/smallstudyeffects.html}{Chapter 9.1}

Egger M, Smith GD, Schneider M & Minder C (1997), Bias in meta-analysis detected by a simple,
graphical test. \emph{BMJ}, 315, 629–634.

Sterne, JAC et al. (2011), Recommendations for Examining and Interpreting Funnel Plot
Asymmetry in Meta-Analyses of Randomised Controlled Trials. \emph{BMJ}
343, 1, doi: 10.1136/bmj.d4002 .
}
\seealso{
\code{\link[meta]{metabias}}
}
\author{
Mathias Harrer & David Daniel Ebert
}
