% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.outliers.R
\name{find.outliers}
\alias{find.outliers}
\alias{spot.outliers.random}
\alias{spot.outliers.fixed}
\alias{spot.outliers}
\title{Find Statistical Outliers in a Meta-Analysis}
\usage{
find.outliers(x, ...)
}
\arguments{
\item{x}{Either (1) an object of class \code{meta}, generated by the \code{metabin}, \code{metagen},
\code{metacont}, \code{metacor}, \code{metainc}, \code{metarate} or \code{metaprop} function; or (2)
and object of class \code{rma.uni} created with the \code{\link[metafor]{rma.uni}} function in \code{metafor}.}

\item{...}{Additional parameters for the \code{\link[metafor]{rma.uni}} or \code{\link[meta]{update.meta}} function.}
}
\value{
Returns the identified outliers and the meta-analysis results when the outliers are removed.

If the provided meta-analysis object is of class \code{meta}, the following objects are returned if the
results of the function are saved to another object:
\itemize{
\item \code{out.study.fixed}: A numeric vector containing the names of the outlying studies when
assuming a fixed-effect model.
\item \code{out.study.random}: A numeric vector containing the names of the outlying studies when
assuming a random-effects model. The \eqn{\tau^{2}} estimator \code{method.tau} is inherited from \code{x}.
\item \code{m.fixed}: An object of class \code{meta} containing the results of the meta-analysis with outliers
removed (assuming a fixed-effect model).
\item \code{m.random}: An object of class \code{meta} containing the results of the meta-analysis with outliers
removed (assuming a random-effects model, and using the same \code{method.tau} as in the original analysis).
}

If the provided meta-analysis object is of class \code{rma.uni}, the following objects are returned if the
results of the function are saved to another object:
\itemize{
\item \code{out.study}: A numeric vector containing the names of the outlying studies.
\item \code{m}: An object of class \code{rma.uni} containing the results of the meta-analysis with outliers
removed (using the same settings as in the meta-analysis object provided).
}
}
\description{
Searches for statistical outliers in meta-analysis results generated by \code{\link[meta]{meta}} functions or the
\code{\link[metafor]{rma.uni}} in the \code{metafor} package.
}
\details{
This function searches for outlying studies in a meta-analysis results object. Studies are defined as outliers when
their 95\% confidence interval lies ouside the 95\% confidence interval of the pooled effect.

When outliers are found, the function automatically recalculates the meta-analysis results, using the same settings as
in the object provided in \code{x}, but excluding the detected outliers.

A forest plot of the meta-analysis with outliers removed can be generated directly by plugging the output of the function into
the \code{forest} function.
}
\examples{
suppressPackageStartupMessages(library(meta))
suppressPackageStartupMessages(library(metafor))
suppressPackageStartupMessages(library(dmetar))

# Pool with meta
m1 <- metagen(TE, seTE, data = ThirdWave,
              studlab = ThirdWave$Author, comb.fixed = FALSE)

# Pool with metafor
m2 <- rma(yi = TE, sei = seTE, data = ThirdWave,
          slab = ThirdWave$Author, method = "PM")

# Find outliers
fo1 <- find.outliers(m1)
fo2 <- find.outliers(m2)

# Show summary
summary(fo1)
summary(fo2)

\dontrun{
# Make forest plot
# Pass additional arguments from meta & metafor's forest function
forest(fo1, prediction = TRUE)
forest(fo2, cex = .8, col = "lightblue")
}
}
\references{
Harrer, M., Cuijpers, P., Furukawa, T.A, & Ebert, D. D. (2019).
\emph{Doing Meta-Analysis in R: A Hands-on Guide}. DOI: 10.5281/zenodo.2551803. \href{https://bookdown.org/MathiasHarrer/Doing_Meta_Analysis_in_R/detecting-outliers-influential-cases.html}{Chapter 6.2}
}
\seealso{
\code{\link[metafor]{influence.rma.uni}}, \code{\link[meta]{metainf}}, \code{\link[meta]{baujat}}
}
\author{
Mathias Harrer & David Daniel Ebert
}
