% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gosh.diagnostics.R
\name{gosh.diagnostics}
\alias{gosh.diagnostics}
\title{Identify studies contributing to heterogeneity patterns found in GOSH plots}
\usage{
gosh.diagnostics(data, km = TRUE, db = TRUE, gmm = TRUE, km.centers =
  3, db.eps = 0.15, db.min.pts = 20, gmm.diag = TRUE, gmm.clusters = 2,
  gmm.tolerance = 1e-16, gmm.itermax = 10000, seed = 123)
}
\arguments{
\item{data}{An object of class \code{gosh.rma} created through the
\code{\link[metafor]{gosh}} function.}

\item{km}{Logical. Should the \emph{k}-Means algorithm be used to identify
patterns in the GOSH plot matrix? \code{TRUE} by default.}

\item{db}{Logical. Should the DBSCAN algorithm be used to identify patterns
in the GOSH plot matrix? \code{TRUE} by default.}

\item{gmm}{Logical. Should a bivariate Gaussian Mixture Model be used to
identify patterns in the GOSH plot matrix? \code{TRUE} by default.}

\item{km.centers}{Number of clusters to assume for the \emph{k}-Means
algorithm. Default is \code{3}.}

\item{db.eps}{The distance radius \eqn{\epsilon} used to determine clusters
(DBSCAN). Default is \code{0.15}.}

\item{db.min.pts}{The minimum number of points \eqn{minPts} required within
\eqn{\epsilon} used to determine clusters. Default is \code{20}.}

\item{gmm.diag}{Logical. Should the covariance matrix of the components be
restricted to diagonal matrices? Default is \code{TRUE}.}

\item{gmm.clusters}{Number of clusters to assume for the Gaussian Mixture
Model. Default is \code{2}.}

\item{gmm.tolerance}{Relative change threshold of log-likelihood used to stop
the Expectation-Maximization algorithm for the Gaussian Mixture Model.
Default is \code{1e-16}.}

\item{gmm.itermax}{Maximum number of iterations for the
Expectation-Maximization algorithm used in the Gaussian Mixture Model to
reach convergence. Default is \code{10000}.}

\item{seed}{Seed used for reproducibility. Default seed is \code{123}.}
}
\description{
This function uses three unsupervised learning learning algorithms
(\emph{k}-means, DBSCAN and Gaussian Mixture Models) to identify studies
contributing to the heterogeneity-effect size patterns found in GOSH (graphic
display of study heterogeneity) plots.
}
\details{
\strong{GOSH Plots}

  GOSH (\emph{graphic display of study
  heterogeneity}) plots were proposed by Olkin, Dahabreh and Trikalinos
  (2012) as a diagnostic plot to assess effect size heterogeneity. GOSH plots
  facilitate the detection of both (i) outliers and (ii) distinct homogeneous
  subgroups within the modeled data.

  Data for the plots is generated by fitting a random-effects-model with the
  same specifications as in the meta-analysis to all \eqn{\mathcal{P}(k),
  \emptyset \notin \mathcal{P}(k), \forall 2^{k-1} \leq 10^6} possible
  subsets of studies in an analysis. For \eqn{|\mathcal{P}(k)| > 10^6}, 1
  million subsets are randomly sampled and used for model fitting when using
  the \code{\link[metafor]{gosh}} function.


  \strong{GOSH Plot Diagnostics}

  Although GOSH plots allow to detect heterogeneity patterns and distinct
  subgroups within the data, interpretation which studies contribute to a
  certain subgroup or pattern is often difficult or computationally
  intensive. To facilitate the detection of studies responsible for specific
  patterns within the GOSH plots, this function randomly samples \eqn{10^4}
  data points from the GOSH Plot data (to speed up computation). Of the data
  points, only the \eqn{z}-transformed \eqn{I^2} and effect size value is
  used (as other heterogeneity metrics produced for the GOSH plot data using
  the \code{\link[metafor]{gosh}} function are linear combinations of
  \eqn{I^2}). To this data, three clustering algorithms are applied.
  \itemize{ \item The first algorithm is \emph{k}-Means clustering using the
  algorithm by Hartigan & Wong (1979) and \eqn{m_k = 3} cluster centers by
  default. The functions uses the \code{\link[stats]{kmeans}} implementation
  to perform \emph{k}-Means clustering. \item As \emph{k}-Means does not
  perform well in the presence of distinct arbitrary subclusters and noise,
  the function also applies \strong{DBSCAN} (\emph{density reachability and
  connectivity clustering}; Schubert et al., 2017). The hyperparameters
  \eqn{\epsilon} and \eqn{MinPts} can be tuned for each analysis to maintain
  a reasonable amount of granularity while not producing too many
  subclusters. The function uses the \code{\link[fpc]{dbscan}} implementation
  to perform the DBSCAN clustering. \item Lastly, as a clustering approach
  using a probabilistic model, Gaussian Mixture Models (GMM; Leisch, 2004)
  are integrated in the function using an internal call to the
  \code{\link[flexmix]{flexmix}} implementation. The GMM implemented here use
  the Expectation-Maximization algorithm for clustering, hyperparameters of
  which can be tuned using the \code{gmm.tolerance} and \code{gmm.itermax}
  parameters. }

  To assess which studies predominantly contribute to a detected cluster,
  the function calculates the cluster imbalance of a specific study using the
  difference between (i) the expected share of subsets containing a specific
  study if the cluster makeup was purely random (viz., representative for the
  full sample), and the (ii) actual share of subsets containing a specific
  study within a cluster. Cook's distance for each study is then calculated
  based on a linear intercept model to determine the leverage of a specific
  study for each cluster makeup. Studies with a leverage value three times
  above the mean in any of the generated clusters (for all used clustering
  algorithms) are returned as potentially influential cases and the GOSH plot
  is redrawn highlighting these specific studies.
}
\examples{
# Example: load gosh data (created with metafor's 'gosh' function),
# then use function
\dontrun{
data("m.gosh")
gosh.diagnostics(m.gosh)
}
}
\references{
Leisch, F. (2004). \emph{Flexmix: A general framework for finite
  mixture models and latent glass regression in R}.

Hartigan, J. A., & Wong, M. A. (1979). Algorithm as 136: A K-Means Clustering Algorithm.
\emph{Journal of the Royal Statistical Society. Series C (Applied Statistics), 28} (1). 100–108.

Olkin, I., Dahabreh, I. J., Trikalinos, T. A. (2012). GOSH–a Graphical Display of Study Heterogeneity.
\emph{Research Synthesis Methods 3}, (3). 214–23.

Schubert, E., Sander, J., Ester, M., Kriegel, H. P. & Xu, X. (2017). DBSCAN Revisited, Revisited:
Why and How You Should (Still) Use DBSCAN. \emph{ACM Transactions on Database Systems (TODS) 42}, (3). ACM: 19.
}
\seealso{
\code{\link{InfluenceAnalysis}}
}
\author{
Mathias Harrer & David Daniel Ebert
}
