% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.gosh.R
\docType{data}
\name{m.gosh}
\alias{m.gosh}
\title{GOSH plot dataset}
\format{
Dataset of class \code{gosh.rma}.
}
\source{
Data generated by fitting a random-effects meta-analysis with \code{rma.uni}
using \code{dmetar}s in-built \code{ThirdWave} dataset, which was then provided as input to the
\code{gosh} function in \code{metafor}.
}
\usage{
data("m.gosh")
}
\description{
GOSH plot dataset
}
\author{
Mathias Harrer, David Daniel Ebert
}
\keyword{datasets}
