% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gosh.diagnostics.R
\name{plot.gosh.diagnostics}
\alias{plot.gosh.diagnostics}
\title{Print, summary and plot methods for objects created using 'dmetar' functions}
\usage{
\method{plot}{gosh.diagnostics}(x, which = c("all", "cluster", "outlier"), ...)
}
\arguments{
\item{x}{An object of class \code{gosh.diagnostics}.}

\item{which}{Type of plot to display. Can be \code{"all"} (all plots), \code{"cluster"}
(cluster plots only) or \code{"outlier"} (outlier plots only, if available).}

\item{...}{Additional arguments.}
}
\description{
Print, summary and plot S3 methods for objects of class \code{direct.evidence.plot}, \code{find.outliers},
\code{influence.analysis}, \code{multimodel.inference}, \code{pcurve}, \code{power.analysis},
\code{subgroup.analysis.mixed.effects}, \code{eggers.test}, \code{gosh.diagnostics}, and \code{sucra}.
}
\details{
A total of four package-specific S3 methods are provided in \code{dmetar}: S3 methods for \code{print},
\code{summary}, \code{plot} and \code{forest}. Outputs from \code{print} and \code{summary} are always identical.
}
\author{
Mathias Harrer & David Daniel Ebert
}
