% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.InfluenceAnalysis.R
\name{plot.InfluenceAnalysis}
\alias{plot.InfluenceAnalysis}
\title{Print, summary and plot methods for objects created using 'dmetar' functions}
\usage{
\method{plot}{InfluenceAnalysis}(x, which = "all", ...)
}
\arguments{
\item{x}{An object of class \code{influence.analysis}.}

\item{which}{Which plot(s) should be generated? Can be either \code{"all"} (default), \code{"baujat"}, \code{"influence"},
\code{"ES"} (forest plot sorted by effect size) or \code{"I2"} (forest plot sorted by heterogeneity).}

\item{...}{Additional arguments.}
}
\description{
Print, summary and plot S3 methods for objects of class \code{direct.evidence.plot}, \code{find.outliers},
\code{influence.analysis}, \code{multimodel.inference}, \code{pcurve}, \code{power.analysis},
\code{subgroup.analysis.mixed.effects}, and \code{sucra}.
}
\details{
A total of four package-specific S3 methods are provided in \code{dmetar}: S3 methods for \code{print},
\code{summary}, \code{plot} and \code{forest}. Outputs from \code{print} and \code{summary} are always identical.
}
\author{
Mathias Harrer & David Daniel Ebert
}
