% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.analysis.subgroup.R
\name{power.analysis.subgroup}
\alias{power.analysis.subgroup}
\title{A priori power calculator for subgroup contrasts}
\usage{
power.analysis.subgroup(TE1, TE2, seTE1, seTE2, sd1, sd2, var1, var2,
    two.tailed=TRUE)
}
\arguments{
\item{TE1}{Pooled effect size (e.g., standardized mean difference, Hedges' \eqn{g}, log-Odds Ratio or other
linear continuous effect size) of the first subgroup of studies.}

\item{TE2}{Pooled effect size (e.g., standardized mean difference, Hedges' \eqn{g}, log-Odds Ratio or other
linear continuous effect size) of the second subgroup of studies.}

\item{seTE1}{Pooled standard error of the first subgroup of studies. Either \code{seTE1/seTE2},
\code{sd1/sd2}, or \code{var1/var2} must be provided.}

\item{seTE2}{Pooled standard error of the second subgroup of studies. Either \code{seTE1/seTE2},
\code{sd1/sd2}, or \code{var1/var2} must be provided.}

\item{sd1}{Pooled standard deviation of the first subgroup of studies. Either \code{seTE1/seTE2},
\code{sd1/sd2}, or \code{var1/var2} must be provided.}

\item{sd2}{Pooled standard deviation of the second subgroup of studies. Either \code{seTE1/seTE2},
\code{sd1/sd2}, or \code{var1/var2} must be provided.}

\item{var1}{Pooled variance of the first subgroup of studies. Either \code{seTE1/seTE2},
\code{sd1/sd2}, or \code{var1/var2} must be provided.}

\item{var2}{Pooled variance of the second subgroup of studies. Either \code{seTE1/seTE2},
\code{sd1/sd2}, or \code{var1/var2} must be provided.}

\item{two.tailed}{Logical. Should a two-tailed (\code{TRUE}) or one-tailed (\code{FALSE}) test (\eqn{\alpha = 0.05}) be assumed?
Default is \code{TRUE}.}
}
\value{
Returns a \code{list} with five elements:
\itemize{
\item \code{Power}: The estimated power of the subgroup contrast, expressed as a value between 0 and 1 (i.e., 0\%-100\%).
\item \code{Plot}: A plot showing the effect size difference (x), power (y), estimated power (red point) and
estimated power for changing effect size differences (blue line). A dashed line at 80\% power is also provided as a
visual threshold for sufficient power.
\item \code{Data}: A \code{data.frame} containing the data used to generate the plot in \code{Plot}.
\item \code{Test}: The type of test used for the power calculations (\code{"one.tailed"} or \code{"two.tailed"}).
\item \code{Gamma}: The analyzed effect size difference calculated from the inputs.
}
}
\description{
This function performs an \emph{a priori} power estimation for a test for subgroup differences
within a meta-analysis.
}
\details{
This function provides an estimate of the power \eqn{1-\beta} of a subgroup contrast analysis provided the
assumed effect sizes in each subgroup and their dispersion measures. The function implements the formulae
described by Hedges and Pigott (2001).
}
\examples{
# Example 1: using standard error and two-tailed test
power.analysis.subgroup(TE1=0.30, TE2=0.66, seTE1=0.13, seTE2=0.14)

# Example 2: using variance and one-tailed test
pasg = power.analysis.subgroup(TE1=-0.91, TE2=-1.22, var1 = 0.0023, var2 = 0.0078,
    two.tailed = FALSE)
summary(pasg)

# Only show plot
plot(pasg)
}
\references{
Hedges, L. V., & Pigott, T. D. (2001). The power of statistical tests in meta-analysis.
\emph{Psychological methods, 6}(3), 203.
}
\seealso{
\code{\link{power.analysis}}
}
\author{
Mathias Harrer & David Daniel Ebert
}
