% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.InfluenceAnalysis.R
\name{print.InfluenceAnalysis}
\alias{print.InfluenceAnalysis}
\title{Print, summary and plot methods for objects created using 'dmetar' functions}
\usage{
\method{print}{InfluenceAnalysis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{InfluenceAnalysis}.}

\item{...}{Additional arguments.}
}
\description{
Print, summary and plot S3 methods for objects of class \code{direct.evidence.plot}, \code{find.outliers},
\code{influence.analysis}, \code{multimodel.inference}, \code{pcurve}, \code{power.analysis},
\code{subgroup.analysis.mixed.effects}, and \code{sucra}.
}
\details{
A total of four package-specific S3 methods are provided in \code{dmetar}: S3 methods for \code{print},
\code{summary}, \code{plot} and \code{forest}. Outputs from \code{print} and \code{summary} are always identical.
}
\author{
Mathias Harrer & David Daniel Ebert
}
