% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroup_analyses_mixed_effects_function.R
\name{subgroup.analysis.mixed.effects}
\alias{subgroup.analysis.mixed.effects}
\alias{sgame}
\title{Subgroup analysis using a mixed-effects model}
\usage{
subgroup.analysis.mixed.effects(x, subgroups, exclude = "none")
}
\arguments{
\item{x}{An object of class \code{meta}, generated by the \code{metabin}, \code{metagen},
\code{metacont}, \code{metacor}, \code{metainc}, or \code{metaprop} function.}

\item{subgroups}{A character vector of the same length as the number of studies within the
meta-analysis, with a unique code for the subgroup each study belongs to. Must have the
same order as the studies in the \code{meta} object.}

\item{exclude}{Single string or concatenated array of strings. The name(s) of the subgroup
levels to be excluded from the subgroup analysis. If \code{"none"} (default), all subgroup
levels are used for the analysis.}
}
\value{
Returns a \code{list} with five objects:
\itemize{
\item \code{within.subgroup.results}: The pooled effect size for each subgroup and corresponding measures of heterogeneity (
\code{Q} and \code{I2}). If the summary measure \code{sm} is defined as one of
\code{"RR"}, \code{"RD"}, \code{"OR"}, \code{"ASD"}, \code{"HR"} or \code{"IRR"} in the
\code{meta} object provided in \code{x}, the backtransformed (exponentiated)
pooled effect for each subgroup effect size along with the 95\% confidence interval is also provided.
\item \code{subgroup.analysis.results}: The results for the \code{Q}-test for subgroup differences, its degrees of freedom \code{df} and
\emph{p}-value.
\item \code{m.random}: An object of class \code{meta} containing the results of the random-effects model applied
for pooling results in each subgroup in the first step.
\item \code{method.tau}: The \eqn{\tau^2} estimator used for within-subgroup pooling
(inherited from the \code{meta} object provided in \code{x}).
\item \code{k}: The total number of included studies.
}
}
\description{
This function performs a mixed-effects (random-effects model within subgroups,
fixed-effect model between subgroups) subgroup analysis using \code{meta} objects.
}
\details{
This function conducts a test for differences in effect sizes between subgroups of a meta-analysis.
The function implements a mixed-effect model, in which the overall effect size for each subgroup is
calculated using a random-effect model, and the test for subgroup differences is conducted using a
fixed-effect model. The implementation follows the fixed-effects (plural) model described in Borenstein
and Higgins (2013).

This model is appropriate for subgroup tests when the subgroup levels under study
are assumed to be exhaustive for the characteristic at hand, and are not randomly chosen instances
of a "population" of subgroup levels. For example, the fixed-effects (plural) model used in the function
is valid when differences between studies published before and after a certain year are considered as a
(binary) subgroup level. When subgroup levels can be assumed to be random samples from a distribution of
subgroup levels, a random-effects model is more appropriate, and may be calculated using
the \code{\link[meta]{update.meta}} function.

The function uses the study effect sizes \code{TE} and their standard error \code{seTE} of the provided
\code{meta} object to perform the subgroup analyses. Specifications of the summary measure \code{sm} are
inherited and used to backtransform log-transformed effect sizes to their original metrics if necessary.

Results can be inspected by plugging the function output into the \code{summary} function. Forest plots
can be generated using \code{forest}. Additional arguments of the \code{\link[meta]{forest.meta}} function
can be passed to the \code{forest} function for additional styling.
}
\examples{
# Example 1: Hedges' g as effect size, precalculated effect sizes
suppressPackageStartupMessages(library(meta))
data("ThirdWave")
ThirdWave = ThirdWave[c(1,2,3,5,9,18),]

m1 <- metagen(TE = TE,
              seTE = seTE,
              studlab = paste(ThirdWave$Author),
              data=ThirdWave,
              comb.fixed = FALSE,
              method.tau = "PM",
              sm = "SMD")

sgame1 = subgroup.analysis.mixed.effects(x = m1, subgroups = ThirdWave$TypeControlGroup)
summary(sgame1)

# Generate Forest Plot
# Additional arguments of the meta::forest.meta can be supplied
forest(sgame1, col.diamond = "darkgreen")


# Example 2: Hedges' g as effect size, raw effect data
suppressPackageStartupMessages(library(meta))
data(amlodipine)

# Create an arbitrary subgroup for illustration purposes
amlodipine$subgroup = rep(c("A","B"),4)

m2 <- metacont(n.amlo, mean.amlo, sqrt(var.amlo),
               n.plac, mean.plac, sqrt(var.plac),
               data=amlodipine, studlab=amlodipine$study,
               sm = "SMD")

sgame2 = subgroup.analysis.mixed.effects(x = m2, subgroups = amlodipine$subgroup)
summary(sgame2)
forest(sgame2)


# Example 3: Risk ratio as effect size, binary outcome data, exlcude one level
suppressPackageStartupMessages(library(meta))
data(Olkin95)

# Create an arbitrary subgroup for illustration purposes
Olkin95$subgroup = c(rep(c("A","B"), 30), rep("C",10))

m3 <- metabin(event.e, n.e, event.c, n.c,
              data = Olkin95, studlab = Olkin95$author,
              method = "Inverse")

# Use shorthand
sgame3 = sgame(x = m3, subgroups = Olkin95$subgroup,
               exclude = "B")
summary(sgame3)


# Example 4: IRR as effect size, incidence data
suppressPackageStartupMessages(library(meta))
data(smoking)

# Create an arbitrary subgroup for illustration purposes
smoking$subgroup = c(rep(c("A"), 4), rep(c("B"), 3))

m4 <- metainc(d.smokers, py.smokers,
              d.nonsmokers, py.nonsmokers,
              data=smoking, studlab=study, sm="IRR")

sgame4 = subgroup.analysis.mixed.effects(x = m4, subgroups = smoking$subgroup)
summary(sgame4)
forest(sgame4)
}
\references{
Harrer, M., Cuijpers, P., Furukawa, T.A, & Ebert, D. D. (2019).
\emph{Doing Meta-Analysis in R: A Hands-on Guide}. DOI: 10.5281/zenodo.2551803. \href{https://bookdown.org/MathiasHarrer/Doing_Meta_Analysis_in_R/subgroup.html}{Chapter 7}.

Borenstein, M. & Higgins, J. P. T. (2013). Meta-Analysis and Subgroups. \emph{Prevention Science, 14} (2): 134–43.
}
\seealso{
\code{\link{multimodel.inference}}
}
\author{
Mathias Harrer & David Daniel Ebert
}
