% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.power.analysis.subgroup.R
\name{summary.power.analysis.subgroup}
\alias{summary.power.analysis.subgroup}
\title{Print, summary and plot methods for objects created using 'dmetar' functions}
\usage{
\method{summary}{power.analysis.subgroup}(object, ...)
}
\arguments{
\item{object}{An object of class \code{power.analysis.subgroup}.}

\item{...}{Additional arguments.}
}
\description{
Print, summary and plot S3 methods for objects of class \code{direct.evidence.plot}, \code{find.outliers},
\code{influence.analysis}, \code{multimodel.inference}, \code{pcurve}, \code{power.analysis},
\code{subgroup.analysis.mixed.effects}, and \code{sucra}.
}
\details{
A total of four package-specific S3 methods are provided in \code{dmetar}: S3 methods for \code{print},
\code{summary}, \code{plot} and \code{forest}. Outputs from \code{print} and \code{summary} are always identical.
}
\author{
Mathias Harrer & David Daniel Ebert
}
