% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thirdwave.R
\docType{data}
\name{ThirdWave}
\alias{ThirdWave}
\title{'Third-Wave' cognitive behavioral interventions for perceived stress in college students dataset}
\format{
A \code{data.frame} with 8 columns.
\describe{
\item{Author}{\code{character}. The study label containing the author(s) of the study.}
\item{TE}{\code{numeric}. The calculated standardized mean difference at post-test between the intervention and control group.}
\item{seTE}{\code{numeric}. The standard error of the standardized mean difference.}
\item{RiskOfBias}{\code{character}. The risk of bias rating according to the Cochrane Risk of Bias Tool.}
\item{TypeControlGroup}{\code{character}. The type of control group used in the study.}
\item{InterventionDuration}{\code{character}. The dichotomized duration of the intervention.}
\item{InterventionType}{\code{character}. The type of third-wave intervention rationale used.}
\item{ModeOfDelivery}{\code{character}. The mode of delivery used for the intervention.}
}
}
\source{
Slightly changed dataset of a meta-analysis on third-wave CBT interventions for
perceived stress in college students.
}
\usage{
data(ThirdWave)
}
\description{
This is a toy dataset containing pre-calculated effect size data of a meta-analysis on
randomized controlled trials comparing the effectiveness of 'third-wave' CBT interventions
for perceived stress in college students to inactive controls. Effect size data is provided
as the standardized mean difference (SMD) between the intervention and control group
and the corresponding standard error for each study at post.
The dataset also contains columns for study characteristics which may serve as potential
effect size moderators.
}
\author{
Mathias Harrer, Eva-Maria Rathner, David Daniel Ebert
}
\keyword{datasets}
