% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.analysis.R
\name{InfluenceAnalysis}
\alias{InfluenceAnalysis}
\title{Influence Diagnostics Plot}
\usage{
InfluenceAnalysis(x, random = FALSE, subplot.heights = c(30,18),
    subplot.widths = c(30,30), forest.lims = 'default',
    return.separate.plots = FALSE, text.scale = 1)
}
\arguments{
\item{x}{An object of class \code{meta}, generated by the \code{metabin}, \code{metagen},
\code{metacont}, \code{metacor}, \code{metainc}, \code{metarate} or \code{metaprop} function.}

\item{random}{Logical. Should the random-effects model be used to generate the influence diagnostics?
Uses the \code{method.tau} specified in the \code{meta} object if one
of "\code{DL}", "\code{HE}", "\code{SJ}", "\code{ML}", "\code{REML}", "\code{EB}", "\code{PM}", "\code{HS}" or "\code{GENQ}" (to ensure compatibility with
the \code{\link[metafor]{metafor}} package). Otherwise, the DerSimonian-Laird
(\code{"DL"}; DerSimonian & Laird, 1986) estimator is used. \code{FALSE} by default.}

\item{subplot.heights}{Concatenated array of two numerics. Specifies the heights of the
first (first number) and second (second number) row of the overall results plot generated by the function.
Default is \code{c(30,18)}.}

\item{subplot.widths}{Concatenated array of two numerics. Specifies the widths of the
first (first number) and second (second number) column of the overall results plot generated by the function.
Default is \code{c(30,30)}.}

\item{forest.lims}{Concatenated array of two numerics. Specifies the x-axis limits of the forest plots
generated by the function. Use \code{"default"} if standard settings should be used (this is the default).}

\item{return.separate.plots}{Logical. Should the influence plots be returned as separate plots in lieu
of returning them in one overall plot? Additionally returns a dataframe \code{Data} containing
the data used for plotting. If set to \code{TRUE}, the output of the function must be saved in a variable;
specific plots can then be accessed by selecting the plot element and using \code{plot()}.}

\item{text.scale}{Positive numeric. Scaling factor for the text geoms used in the plot. Values <1 shrink the
text, while values >1 increase the text size. Default is \code{1}.}
}
\value{
If \code{return.separate.plots} is set to \code{TRUE}, a \code{list} object containing the
following objects is returned (if results are saved to a variable):
\itemize{
\item \code{BaujatPlot}: The Baujat plot
\item \code{InfluenceCharacteristics}: The Viechtbauer-Cheung influence characteristics plot
\item \code{ForestEffectSize}: The forest plot sorted by effect size
\item \code{ForestI2}: The forest plot sorted by between-study heterogeneity
\item \code{Data}: A \code{data.frame} containing the data used for plotting.
}
}
\description{
Conducts an influence analysis of meta-analysis results generated by \code{\link[meta]{meta}} functions
and produces influence diagnostic plots.
}
\details{
The function conducts an influence analysis using the "Leave-One-Out" paradigm internally
and produces four diagnostic plots. These plots may be used to determine which study or effect size
may have an excessive influence on the overall results of a meta-analysis and/or contribute substantially to
the between-study heterogeneity in an analysis. This may be used for outlier detection and to test
the robustness of the overall results found in an analysis. Four plots are produced,
left to right, top to bottom:
\itemize{
\item \strong{Baujat Plot}: Baujat et al. (2002) proposed a plot to evaluate heterogeneity patterns in
a meta-analysis. The x-axis of the Baujat plot shows the overall heterogeneity contribution of each effect size
while the y-axis shows the influence of each effect size on the pooled result. The \code{\link[meta]{metainf}}
function is called internally to produce data used for plotting. Effect sizes or studies with high values
on both the x and y-axis may be considered to be influential cases; effect sizes or studies
with high heterogeneity contribution (x-axis) and low influence on the overall results can be outliers
which might be deleted to reduce the amount of between-study heterogeneity.
\item \strong{Influence Characteristics}: Produces plots for several influence analysis diagnostics
proposed by Viechtbauer & Cheung (2010). Data for plotting are generated by an internal call
to \code{\link[metafor]{influence.rma.uni}}. Effect sizes/studies determined to be influential cases
using the "rules of thumb" described in Viechtbauer & Cheung (2010) are plotted in red. For further
details, see the documentation of the \code{\link[metafor]{influence.rma.uni}} function.
\item \strong{Forest Plot for the Leave-One-Out Analysis, sorted by effect size}: This forest plot
displays the effect size and \eqn{I^2}-heterogeneity when omitting one of the \eqn{k} studies each time.
The plot is sorted by effect size to determine which studies or effect sizes particularly
affect the overall effect size. Results are generated by an internal call to \code{\link[meta]{metainf}}.
\item \strong{Forest Plot for the Leave-One-Out Analysis, sorted by \eqn{I^2}}: see above; results are sorted
by \eqn{I^2} to determine the study for which exclusion results in the greatest reduction of heterogeneity.
}
}
\examples{
# Load 'ThirdWave' data
data(ThirdWave)

# Create 'meta' meta-analysis object
suppressPackageStartupMessages(library(meta))
meta = metagen(TE, seTE, studlab = paste(ThirdWave$Author), data=ThirdWave)

# Return separate plots
InfluenceAnalysis(meta, subplot.heights = c(30,24), text.scale=0.5)
}
\references{
Harrer, M., Cuijpers, P., Furukawa, T.A, & Ebert, D. D. (2019).
\emph{Doing Meta-Analysis in R: A Hands-on Guide}. DOI: 10.5281/zenodo.2551803. \href{https://bookdown.org/MathiasHarrer/Doing_Meta_Analysis_in_R/influenceanalyses.html}{Chapter 6.3}

DerSimonian R. & Laird N. (1986), Meta-analysis in clinical trials. \emph{Controlled Clinical Trials, 7}, 177–188.

Viechtbauer, W., & Cheung, M. W.-L. (2010). Outlier and influence diagnostics for meta-analysis. \emph{Research Synthesis Methods, 1}, 112–125.
}
\seealso{
\code{\link[metafor]{influence.rma.uni}}, \code{\link[meta]{metainf}}, \code{\link[meta]{baujat}}
}
\author{
Mathias Harrer & David Daniel Ebert
}
